+++ coffeenet-autoconfigure/src/test/java/coffee/synyx/autoconfigure/discovery/service/DevelopmentCoffeeNetAppServiceTest.java
@@ -4,14 +4,15 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.util.List;
-
-import static org.hamcrest.CoreMatchers.is;
+import java.util.Map;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.emptyIterable;
+import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasSize;
+
+import static org.hamcrest.core.Is.is;
 
 
 /**
@@ -31,13 +32,27 @@ public class DevelopmentCoffeeNetAppServiceTest {
     @Test
     public void getApps() {
 
-        List<CoffeeNetApp> coffeeNetApps = sut.getApps();
-        assertThat(coffeeNetApps, hasSize(3));
-        assertThat(coffeeNetApps.get(0).getName(), is("Coffee App"));
-        assertThat(coffeeNetApps.get(0).getAuthorities(), is(emptyIterable()));
-        assertThat(coffeeNetApps.get(1).getName(), is("Profile"));
-        assertThat(coffeeNetApps.get(1).getAuthorities(), is(emptyIterable()));
-        assertThat(coffeeNetApps.get(2).getName(), is("CoffeeNet Admin App"));
-        assertThat(coffeeNetApps.get(2).getAuthorities(), contains("ROLE_COFFEENET-ADMIN"));
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = sut.getApps();
+        assertThat(coffeeNetApps.entrySet(), hasSize(3));
+        assertThat(coffeeNetApps.get("Coffee App").get(0).getName(), is("Coffee App"));
+        assertThat(coffeeNetApps.get("Coffee App").get(0).getAuthorities(), is(emptyIterable()));
+        assertThat(coffeeNetApps.get("Profile").get(0).getName(), is("Profile"));
+        assertThat(coffeeNetApps.get("Profile").get(0).getAuthorities(), is(emptyIterable()));
+        assertThat(coffeeNetApps.get("Coffee Admin App").get(0).getName(), is("Coffee Admin App"));
+        assertThat(coffeeNetApps.get("Coffee Admin App").get(0).getAuthorities(), hasItem("ROLE_COFFEENET-ADMIN"));
+    }
+
+
+    @Test
+    public void getAppsWithQuery() {
+
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = sut.getApps(AppQuery.builder().build());
+        assertThat(coffeeNetApps.entrySet(), hasSize(3));
+        assertThat(coffeeNetApps.get("Coffee App").get(0).getName(), is("Coffee App"));
+        assertThat(coffeeNetApps.get("Coffee App").get(0).getAuthorities(), is(emptyIterable()));
+        assertThat(coffeeNetApps.get("Profile").get(0).getName(), is("Profile"));
+        assertThat(coffeeNetApps.get("Profile").get(0).getAuthorities(), is(emptyIterable()));
+        assertThat(coffeeNetApps.get("Coffee Admin App").get(0).getName(), is("Coffee Admin App"));
+        assertThat(coffeeNetApps.get("Coffee Admin App").get(0).getAuthorities(), hasItem("ROLE_COFFEENET-ADMIN"));
     }
 }