+++ coffeenet-autoconfigure/src/test/java/coffee/synyx/autoconfigure/discovery/service/IntegrationCoffeeNetAppServiceTest.java
@@ -15,7 +15,10 @@ import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.cloud.netflix.eureka.EurekaDiscoveryClient.EurekaServiceInstance;
 
 import java.util.List;
+import java.util.Map;
 
+import static org.hamcrest.Matchers.empty;
+import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasSize;
 
 import static org.hamcrest.core.Is.is;
@@ -26,6 +29,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 
 
@@ -35,7 +39,7 @@ import static java.util.Collections.singletonList;
 @RunWith(MockitoJUnitRunner.class)
 public class IntegrationCoffeeNetAppServiceTest {
 
-    private IntegrationCoffeeNetAppService eurekaAppService;
+    private IntegrationCoffeeNetAppService coffeeNetAppService;
 
     @Mock
     private DiscoveryClient discoveryClientMock;
@@ -43,44 +47,175 @@ public class IntegrationCoffeeNetAppServiceTest {
     @Before
     public void setUp() {
 
-        eurekaAppService = new IntegrationCoffeeNetAppService(discoveryClientMock);
+        coffeeNetAppService = new IntegrationCoffeeNetAppService(discoveryClientMock);
     }
 
 
     @Test
     public void getApps() {
 
-        when(discoveryClientMock.getServices()).thenReturn(asList("Profile", "Backpage", "Auth"));
+        String profileName = "Profile";
+        EurekaServiceInstance profile = getServiceInstance(profileName, "ROLE_ADMIN, ROLE_USER");
+        EurekaServiceInstance profile2 = getServiceInstance(profileName, "ROLE_USER");
+        when(discoveryClientMock.getInstances(profileName)).thenReturn(asList(profile, profile2));
 
-        defineEurekaInstance("Profile");
-        defineEurekaInstance("Backpage");
-        defineEurekaInstance("Auth");
+        String backPageName = "Backpage";
+        EurekaServiceInstance backPage = getServiceInstance(backPageName, null);
+        when(discoveryClientMock.getInstances(backPageName)).thenReturn(singletonList(backPage));
 
-        List<CoffeeNetApp> coffeeNetApps = eurekaAppService.getApps();
-        assertThat(coffeeNetApps, hasSize(3));
-        assertThat(coffeeNetApps.get(0).getName(), is("Auth"));
-        assertThat(coffeeNetApps.get(1).getName(), is("Backpage"));
-        assertThat(coffeeNetApps.get(2).getName(), is("Profile"));
+        String authName = "Auth";
+        EurekaServiceInstance auth = getServiceInstance(authName, "ROLE_USER");
+        when(discoveryClientMock.getInstances(authName)).thenReturn(singletonList(auth));
+
+        when(discoveryClientMock.getServices()).thenReturn(asList(profileName, backPageName, authName));
+
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = coffeeNetAppService.getApps();
+        assertThat(coffeeNetApps.entrySet(), hasSize(3));
+        assertThat(coffeeNetApps.get(profileName), hasSize(2));
+        assertThat(coffeeNetApps.get(profileName).get(0).getName(), is(profileName));
+        assertThat(coffeeNetApps.get(profileName).get(0).getAuthorities(), hasSize(2));
+        assertThat(coffeeNetApps.get(profileName).get(0).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(profileName).get(0).getAuthorities(), hasItem("ROLE_ADMIN"));
+        assertThat(coffeeNetApps.get(profileName).get(1).getName(), is(profileName));
+        assertThat(coffeeNetApps.get(profileName).get(1).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(backPageName), hasSize(1));
+        assertThat(coffeeNetApps.get(backPageName).get(0).getName(), is(backPageName));
+        assertThat(coffeeNetApps.get(backPageName).get(0).getAuthorities(), is(empty()));
+        assertThat(coffeeNetApps.get(authName), hasSize(1));
+        assertThat(coffeeNetApps.get(authName).get(0).getName(), is(authName));
+        assertThat(coffeeNetApps.get(authName).get(0).getAuthorities(), hasSize(1));
+        assertThat(coffeeNetApps.get(authName).get(0).getAuthorities(), hasItem("ROLE_USER"));
+    }
+
+
+    @Test
+    public void getAppsFilteredByAppNamesAndRoles() {
+
+        String frontPageName = "Frontpage";
+        EurekaServiceInstance frontPage = getServiceInstance(frontPageName, null);
+        EurekaServiceInstance frontPageWithUser = getServiceInstance(frontPageName, "ROLE_USER");
+        EurekaServiceInstance frontPageWithAdminAndUser = getServiceInstance(frontPageName, "ROLE_USER, ROLE_ADMIN");
+        when(discoveryClientMock.getInstances(frontPageName)).thenReturn(asList(frontPageWithUser,
+                frontPageWithAdminAndUser, frontPage));
+
+        String backPageName = "Backpage";
+        EurekaServiceInstance backPage = getServiceInstance(backPageName, "ROLE_USER, ROLE_ADMIN");
+        when(discoveryClientMock.getInstances(backPageName)).thenReturn(singletonList(backPage));
+
+        when(discoveryClientMock.getServices()).thenReturn(asList(frontPageName, backPageName));
+
+        AppQuery query = AppQuery.builder().withAppName(frontPageName).withRole("ROLE_USER").build();
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = coffeeNetAppService.getApps(query);
+        assertThat(coffeeNetApps.entrySet(), hasSize(1));
+        assertThat(coffeeNetApps.get(frontPageName), hasSize(3));
+        assertThat(coffeeNetApps.get(frontPageName).get(0).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(0).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName).get(1).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(1).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName).get(2).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(2).getAuthorities(), is(empty()));
+    }
+
+
+    @Test
+    public void getAppsFilteredByRoles() {
+
+        String frontPageName = "Frontpage";
+        EurekaServiceInstance frontPage = getServiceInstance(frontPageName, null);
+        EurekaServiceInstance frontPageWithUser = getServiceInstance(frontPageName, "ROLE_USER");
+        EurekaServiceInstance frontPageWithAdminAndUser = getServiceInstance(frontPageName, "ROLE_USER, ROLE_ADMIN");
+        when(discoveryClientMock.getInstances(frontPageName)).thenReturn(asList(frontPageWithUser,
+                frontPageWithAdminAndUser, frontPage));
+
+        String backPageName = "Backpage";
+        EurekaServiceInstance backPage = getServiceInstance(backPageName, "ROLE_USER, ROLE_ADMIN");
+        when(discoveryClientMock.getInstances(backPageName)).thenReturn(singletonList(backPage));
+
+        when(discoveryClientMock.getServices()).thenReturn(asList(frontPageName, backPageName));
+
+        AppQuery query = AppQuery.builder().withRole("ROLE_USER").build();
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = coffeeNetAppService.getApps(query);
+        assertThat(coffeeNetApps.entrySet(), hasSize(2));
+        assertThat(coffeeNetApps.get(backPageName), hasSize(1));
+        assertThat(coffeeNetApps.get(backPageName).get(0).getName(), is(backPageName));
+        assertThat(coffeeNetApps.get(backPageName).get(0).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName), hasSize(3));
+        assertThat(coffeeNetApps.get(frontPageName).get(0).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(0).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName).get(1).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(1).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName).get(2).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(2).getAuthorities(), is(empty()));
+    }
+
+
+    @Test
+    public void getAppsFilteredByNames() {
+
+        String frontPageName = "Frontpage";
+        EurekaServiceInstance frontPage = getServiceInstance(frontPageName, null);
+        EurekaServiceInstance frontPageWithUser = getServiceInstance(frontPageName, "ROLE_USER");
+        EurekaServiceInstance frontPageWithAdminAndUser = getServiceInstance(frontPageName, "ROLE_USER, ROLE_ADMIN");
+        when(discoveryClientMock.getInstances(frontPageName)).thenReturn(asList(frontPageWithUser,
+                frontPageWithAdminAndUser, frontPage));
+
+        String backPageName = "Backpage";
+        EurekaServiceInstance backPage = getServiceInstance(backPageName, "ROLE_USER, ROLE_ADMIN");
+        when(discoveryClientMock.getInstances(backPageName)).thenReturn(singletonList(backPage));
+
+        when(discoveryClientMock.getServices()).thenReturn(asList(frontPageName, backPageName));
+
+        AppQuery query = AppQuery.builder().withAppName("Frontpage").build();
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = coffeeNetAppService.getApps(query);
+        assertThat(coffeeNetApps.entrySet(), hasSize(1));
+        assertThat(coffeeNetApps.get(frontPageName), hasSize(3));
+        assertThat(coffeeNetApps.get(frontPageName).get(0).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(0).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName).get(1).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(1).getAuthorities(), hasItem("ROLE_USER"));
+        assertThat(coffeeNetApps.get(frontPageName).get(2).getName(), is(frontPageName));
+        assertThat(coffeeNetApps.get(frontPageName).get(2).getAuthorities(), is(empty()));
+    }
+
+
+    @Test
+    public void getAppsFilteredByCaseInsensitiveNames() {
+
+        String frontPageName = "Frontpage";
+        String frontPageNameCI = "frontpage";
+        EurekaServiceInstance frontPageWithAdminAndUser = getServiceInstance(frontPageName, null);
+        when(discoveryClientMock.getInstances(frontPageName)).thenReturn(singletonList(frontPageWithAdminAndUser));
+
+        when(discoveryClientMock.getServices()).thenReturn(singletonList(frontPageName));
+
+        AppQuery query = AppQuery.builder().withAppName(frontPageNameCI).build();
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = coffeeNetAppService.getApps(query);
+        assertThat(coffeeNetApps.entrySet(), hasSize(1));
+        assertThat(coffeeNetApps.get(frontPageNameCI).get(0).getName(), is(frontPageName));
     }
 
 
     @Test
     public void getAppsAndInstancesNotFound() {
 
-        when(discoveryClientMock.getServices()).thenReturn(asList("Frontpage", "Backpage"));
+        when(discoveryClientMock.getServices()).thenReturn(emptyList());
 
-        List<CoffeeNetApp> coffeeNetApps = eurekaAppService.getApps();
-        assertThat(coffeeNetApps, hasSize(0));
+        Map<String, List<CoffeeNetApp>> coffeeNetApps = coffeeNetAppService.getApps();
+        assertThat(coffeeNetApps.entrySet(), hasSize(0));
     }
 
 
-    private void defineEurekaInstance(String name) {
+    private EurekaServiceInstance getServiceInstance(String name, String roles) {
 
         InstanceInfo instanceInfo = InstanceInfo.Builder.newBuilder().setAppName(name).setVIPAddress(name).build();
-        instanceInfo.getMetadata().put("allowedAuthorities", "ROLE_ADMIN");
+
+        if (roles != null) {
+            instanceInfo.getMetadata().put("allowedAuthorities", roles);
+        }
 
         EurekaServiceInstance eurekaServiceInstance = mock(EurekaServiceInstance.class);
         when(eurekaServiceInstance.getInstanceInfo()).thenReturn(instanceInfo);
-        when(discoveryClientMock.getInstances(name)).thenReturn(singletonList(eurekaServiceInstance));
+
+        return eurekaServiceInstance;
     }
 }