+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ### 0.16.0
 * Upgrade to spring boot 1.5.1
+* Add `CoffeeNetAppService` can be queried to find the `CoffeeNetApp` you need
 * Fix get `CoffeeNetApps` in discovery starter ordered by their names
 * Fix hamburger menu in thymeleaf starter
 * Remove bootstrap dependency from navigation bar
 * Remove deprecated `coffeenet/user` and `coffeenet/apps` endpoints for the javascript navigation bar
+
 
 ### 0.15.1
 * Fix to only configure web starters when security and discovery is available